+++ Assets/OxGFrame/AssetLoader/Scripts/Runtime/Cacher/CacheResource.cs
@@ -47,8 +47,12 @@ namespace OxGFrame.AssetLoader.Cacher
                 // 如果有進行 Loading 標記後, 直接 return
                 if (this.HasInLoadingFlags(assetName) && !this.GetRetryCounter(assetName).IsRetryValid())
                 {
-                    this.RemoveLoadingFlags(assetName);
-                    Logging.Print<Logger>($"<color=#FFDC8A>asset: {assetName} Load failed and cannot retry anymore!!! Please to checke asset is existing.</color>");
+                    if (this.GetRetryCounter(assetName).IsRetryActive())
+                    {
+                        this.RemoveLoadingFlags(assetName);
+                        Logging.Print<Logger>($"<color=#ff9b3e>Asset: {assetName} Load failed and cannot retry anymore!!! Please to check asset is existing.</color>");
+                    }
+                    else Logging.Print<Logger>($"<color=#ff9b3e>Asset: {assetName} Loading...</color>");
                     return;
                 }
 
@@ -61,6 +65,7 @@ namespace OxGFrame.AssetLoader.Cacher
                     this.currentCount++;
                     progression?.Invoke(this.currentCount / this.totalCount, this.currentCount, this.totalCount);
                     this.RemoveLoadingFlags(assetName);
+                    Logging.Print<Logger>($"<color=#fff6ba>【Preload】 => Current << CacheResource >> Cache Count: {this.Count}, asset: [{assetName}] already preloaded!!!</color>");
                     continue;
                 }
 
@@ -101,8 +106,9 @@ namespace OxGFrame.AssetLoader.Cacher
                 }
                 else
                 {
+                    if (this.GetRetryCounter(assetName).IsRetryActive()) Logging.Print<Logger>($"<color=#f7ff3e>【Preload】 => << CacheResource >> Asset: {assetName} doing retry. Retry count: {this.GetRetryCounter(assetName).retryCount}, Max retry count: {maxRetryCount}</color>");
+                    else Logging.Print<Logger>($"<color=#f7ff3e>【Preload】 => << CacheResource >> Asset: {assetName} start doing retry. Max retry count: {maxRetryCount}</color>");
                     this.GetRetryCounter(assetName).AddRetryCount();
-                    Logging.Print<Logger>($"<color=#ff9600>【Preload】 => << CacheResource >> Asset: {assetName} doing retry. Retry count: {this.GetRetryCounter(assetName).retryCount}, Max retry count: {maxRetryCount}</color>");
                     await this.PreloadAssetAsync<T>(new string[] { assetName }, progression, maxRetryCount);
                     continue;
                 }
@@ -129,8 +135,12 @@ namespace OxGFrame.AssetLoader.Cacher
                 // 如果有進行 Loading 標記後, 直接 return
                 if (this.HasInLoadingFlags(assetName) && !this.GetRetryCounter(assetName).IsRetryValid())
                 {
-                    this.RemoveLoadingFlags(assetName);
-                    Logging.Print<Logger>($"<color=#FFDC8A>asset: {assetName} Load failed and cannot retry anymore!!! Please to checke asset is existing.</color>");
+                    if (this.GetRetryCounter(assetName).IsRetryActive())
+                    {
+                        this.RemoveLoadingFlags(assetName);
+                        Logging.Print<Logger>($"<color=#ff9b3e>Asset: {assetName} Load failed and cannot retry anymore!!! Please to check asset is existing.</color>");
+                    }
+                    else Logging.Print<Logger>($"<color=#ff9b3e>Asset: {assetName} Loading...</color>");
                     return;
                 }
 
@@ -143,6 +153,7 @@ namespace OxGFrame.AssetLoader.Cacher
                     this.currentCount++;
                     progression?.Invoke(this.currentCount / this.totalCount, this.currentCount, this.totalCount);
                     this.RemoveLoadingFlags(assetName);
+                    Logging.Print<Logger>($"<color=#fff6ba>【Preload】 => Current << CacheResource >> Cache Count: {this.Count}, asset: [{assetName}] already preloaded!!!</color>");
                     continue;
                 }
 
@@ -170,8 +181,9 @@ namespace OxGFrame.AssetLoader.Cacher
                 }
                 else
                 {
+                    if (this.GetRetryCounter(assetName).IsRetryActive()) Logging.Print<Logger>($"<color=#f7ff3e>【Preload】 => << CacheResource >> Asset: {assetName} doing retry. Retry count: {this.GetRetryCounter(assetName).retryCount}, Max retry count: {maxRetryCount}</color>");
+                    else Logging.Print<Logger>($"<color=#f7ff3e>【Preload】 => << CacheResource >> Asset: {assetName} start doing retry. Max retry count: {maxRetryCount}</color>");
                     this.GetRetryCounter(assetName).AddRetryCount();
-                    Logging.Print<Logger>($"<color=#ff9600>【Preload】 => << CacheResource >> Asset: {assetName} doing retry. Retry count: {this.GetRetryCounter(assetName).retryCount}, Max retry count: {maxRetryCount}</color>");
                     this.PreloadAsset<T>(new string[] { assetName }, progression, maxRetryCount);
                     continue;
                 }
@@ -188,8 +200,12 @@ namespace OxGFrame.AssetLoader.Cacher
             // 如果有進行 Loading 標記後, 直接 return
             if (this.HasInLoadingFlags(assetName) && !this.GetRetryCounter(assetName).IsRetryValid())
             {
-                this.RemoveLoadingFlags(assetName);
-                Logging.Print<Logger>($"<color=#FFDC8A>asset: {assetName} Load failed and cannot retry anymore!!! Please to checke asset is existing.</color>");
+                if (this.GetRetryCounter(assetName).IsRetryActive())
+                {
+                    this.RemoveLoadingFlags(assetName);
+                    Logging.Print<Logger>($"<color=#ff9b3e>Asset: {assetName} Load failed and cannot retry anymore!!! Please to check asset is existing.</color>");
+                }
+                else Logging.Print<Logger>($"<color=#ff9b3e>Asset: {assetName} Loading...</color>");
                 return null;
             }
 
@@ -253,8 +269,9 @@ namespace OxGFrame.AssetLoader.Cacher
             else
             {
                 this.UnloadAsset(assetName, true);
+                if (this.GetRetryCounter(assetName).IsRetryActive()) Logging.Print<Logger>($"<color=#f7ff3e>【Load】 => << CacheResource >> Asset: {assetName} doing retry. Retry count: {this.GetRetryCounter(assetName).retryCount}, Max retry count: {maxRetryCount}</color>");
+                else Logging.Print<Logger>($"<color=#f7ff3e>【Load】 => << CacheResource >> Asset: {assetName} start doing retry. Max retry count: {maxRetryCount}</color>");
                 this.GetRetryCounter(assetName).AddRetryCount();
-                Logging.Print<Logger>($"<color=#90FF71>【Load】 => << CacheResource >> Asset: {assetName} doing retry. Retry count: {this.GetRetryCounter(assetName).retryCount}, Max retry count: {maxRetryCount}</color>");
                 return await this.LoadAssetAsync<T>(assetName, progression, maxRetryCount);
             }
 
@@ -270,8 +287,12 @@ namespace OxGFrame.AssetLoader.Cacher
             // 如果有進行 Loading 標記後, 直接 return
             if (this.HasInLoadingFlags(assetName) && !this.GetRetryCounter(assetName).IsRetryValid())
             {
-                this.RemoveLoadingFlags(assetName);
-                Logging.Print<Logger>($"<color=#FFDC8A>asset: {assetName} Load failed and cannot retry anymore!!! Please to checke asset is existing.</color>");
+                if (this.GetRetryCounter(assetName).IsRetryActive())
+                {
+                    this.RemoveLoadingFlags(assetName);
+                    Logging.Print<Logger>($"<color=#ff9b3e>Asset: {assetName} Load failed and cannot retry anymore!!! Please to check asset is existing.</color>");
+                }
+                else Logging.Print<Logger>($"<color=#ff9b3e>Asset: {assetName} Loading...</color>");
                 return null;
             }
 
@@ -323,8 +344,9 @@ namespace OxGFrame.AssetLoader.Cacher
             else
             {
                 this.UnloadAsset(assetName, true);
+                if (this.GetRetryCounter(assetName).IsRetryActive()) Logging.Print<Logger>($"<color=#f7ff3e>【Load】 => << CacheResource >> Asset: {assetName} doing retry. Retry count: {this.GetRetryCounter(assetName).retryCount}, Max retry count: {maxRetryCount}</color>");
+                else Logging.Print<Logger>($"<color=#f7ff3e>【Load】 => << CacheResource >> Asset: {assetName} start doing retry. Max retry count: {maxRetryCount}</color>");
                 this.GetRetryCounter(assetName).AddRetryCount();
-                Logging.Print<Logger>($"<color=#90FF71>【Load】 => << CacheResource >> Asset: {assetName} doing retry. Retry count: {this.GetRetryCounter(assetName).retryCount}, Max retry count: {maxRetryCount}</color>");
                 return this.LoadAsset<T>(assetName, progression, maxRetryCount);
             }
 
@@ -339,7 +361,7 @@ namespace OxGFrame.AssetLoader.Cacher
 
             if (this.HasInLoadingFlags(assetName) && !this.GetRetryCounter(assetName).IsRetryValid())
             {
-                Logging.Print<Logger>($"<color=#FFDC8A>asset: {assetName} Loading...</color>");
+                Logging.Print<Logger>($"<color=#ff9b3e>Asset: {assetName} Loading...</color>");
                 return;
             }